feat: Add new composable useRecognitionFlow for better state management in recognition flow
- Improved the state management of the recognition flow by introducing a new composable `useRecognitionFlow`.
- Added types for screen states and functions to submit audio, handle micro click and reset app.
