+++ .gitignore
@@ -2,6 +2,7 @@
 .output
 .data
 .nuxt
+.nuxt-build
 .nitro
 .cache
 dist
